requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-vision==3.4.4
pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
reportlab==4.0.7
python-dotenv==1.0.0
werkzeug==3.0.1
pymupdf==1.23.8
openai==0.28.1
gunicorn==21.2.0
twilio==8.10.0
sendgrid==6.10.0
pypdf==3.17.4
cachetools==5.3.2
pdfplumber==0.11.0